+++ app/src/main/java/fc/be/app/domain/review/repository/ReviewRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 
-    List<Review> findByPlaceId(Long placeId);
+    List<Review> findByPlaceId(Integer placeId);
 
     List<Review> findByMemberId(Long memberId);
 }
